# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  tag.CreateTagRequest:
    properties:
      name:
        type: string
    type: object
  tag.DeleteTagResponse:
    properties:
      message:
        type: string
    type: object
  tag.FamousTag:
    properties:
      count:
        type: integer
      name:
        type: string
    type: object
  tag.GetAllTagsResponse:
    properties:
      tags:
        items:
          $ref: '#/definitions/tag.Tag'
        type: array
    type: object
  tag.GetFamousTagsRes:
    properties:
      tags:
        items:
          $ref: '#/definitions/tag.FamousTag'
        type: array
    type: object
  tag.Tag:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        description: UUID
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  tag.UpdateTagRequest:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a forum API gateway service.
  title: Forum API Gateway
  version: "1.0"
paths:
...
  /tags:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all tags with optional pagination
      parameters:
      - description: Find by tags name
        in: query
        name: name
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of tags per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.GetAllTagsResponse'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all tags
      tags:
      - tag
    post:
      consumes:
      - application/json
      description: Create a new tag with given information
      parameters:
      - description: Tag information
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/tag.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tag.Tag'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new tag
      tags:
      - tag
  /tags/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a tag by its unique ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.DeleteTagResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Tag item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a tag by its ID
      tags:
      - tag
    get:
      consumes:
      - application/json
      description: Retrieve a tag by its unique ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.Tag'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Tag item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a tag by its ID
      tags:
      - tag
    put:
      consumes:
      - application/json
      description: Update a tag with the given information
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      - description: Tag information
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/tag.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.Tag'
        "400":
          description: Invalid request body
          schema:
            type: string
        "404":
          description: Tag item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a tag by its ID
      tags:
      - tag
  /tags/popular:
    get:
      consumes:
      - application/json
      description: Retrieve a list of famous tags (most used) with optional pagination
        and sorting.
      parameters:
      - description: Filter by tag name
        in: query
        name: name
        type: string
      - description: Sort by count in descending order
        in: query
        name: desc
        type: boolean
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of tags per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tag.GetFamousTagsRes'
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get famous tags
      tags:
      - tag
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"